Update transform_create_populate_tables1.sql
As per Observation 151 and the ticket DCES-612, there are 3 defendant records which are incorrectly created by Marston (confirmed by them) that we should be ignoring from data migration. DELETE statement added at the end of the script for that. It is hardcoded as we want to catch if this happens again (new duplication) which shouldn't.