Update transform_create_populate_tables1.sql
As per DCES-632, this change is to populate missing clientdefaulterreference fields on casedetails table. The base lookup table will need to be refreshed (from a refresh maat extract) each time before we run this script. But this time it fixes 110 missing records. These are caused by business' manual case creation process, which is informed to them and Joanna is looking at resolving it (with a new excel template).